+++ web/package.json
@@ -22,7 +22,8 @@
     "class-variance-authority": "^0.4.0",
     "react": "17.0.2",
     "react-dom": "17.0.2",
-    "tailwind-merge": "^1.9.1"
+    "tailwind-merge": "^1.9.1",
+    "zod": "^3.20.6"
   },
   "devDependencies": {
     "@faker-js/faker": "^7.6.0",
