+++ web/src/modules/invoices/components/InvoiceListItem/InvoiceListItem.tsx
@@ -3,11 +3,11 @@ import { InvoiceStatus } from 'types/graphql'
 
 import { Link, routes } from '@redwoodjs/router'
 
-import { getInvoiceNumberFromId } from 'src/modules/invoices/utils/get-invoice-number-from-id'
 import { testIds } from 'src/test/ids'
 import { formatDate, formatMoney } from 'src/utils'
 
 import Chevron from '../../assets/chevron.svg'
+import { getInvoiceNumberFromId } from '../../utils'
 import InvoiceStatusBadge from '../InvoiceStatusBadge'
 
 interface IInvoiceListItemProps {
