+++ web/src/modules/invoices/utils/index.ts
@@ -1,0 +1,1 @@
+export * from './get-invoice-number-from-id'
